HHH-8312 - named parameters binding are not correct when used within subquery

+++ hibernate-core/src/main/java/org/hibernate/loader/custom/CustomLoader.java
@@ -459,7 +459,7 @@ public class CustomLoader extends Loader {
 			);
 		}
 		if ( loc instanceof Integer ) {
-			return new int[] { ( ( Integer ) loc ).intValue() };
+			return new int[] { (Integer) loc };
 		}
 		else {
 			return ArrayHelper.toIntArray( ( List ) loc );
